create base class for geocms dataitems
